py/PasswordChecker.py
def validate_password(password):
    # Check if password length is between 8 and 20 characters
    if len(password) < 8 or len(password) > 20:
        return False
    
    # Check if password contains at least one uppercase letter
    if not any(char.isupper() for char in password):
        return False
    
    # Check if password contains at least one lowercase letter
    if not any(char.islower() for char in password):
        return False
    
    # Check if password contains at least one digit
    if not any(char.isdigit() for char in password):
        return False
    
    # Check if password contains at least one special character
    if not any(char in '!@#$%^&*()-_=+[]{}|\\;:\'",.<>/?`~' for char in password):
        return False
    
    # If all checks pass, return True
    return True
